chore: rename domains/{sentry,Sentry}
